+++ cmd/main.go
@@ -3,11 +3,13 @@ package main
 import (
 	"fmt"
 
+	"github.com/caarlos0/gohome"
 	"github.com/getantibody/antibody"
 )
 
 func main() {
 	sh, err := antibody.New(
+		gohome.Cache("__antibody")+"/",
 		[]string{
 			"caarlos0/ports kind:path",
 			"caarlos0/jvm version:gh-pages",
