+++ project/local.go
@@ -1,0 +1,21 @@
+package project
+
+func NewLocal(folder string) Project {
+	return localProject{folder}
+}
+
+type localProject struct {
+	folder string
+}
+
+func (l localProject) Download() error {
+	return nil
+}
+
+func (l localProject) Update() error {
+	return nil
+}
+
+func (l localProject) Folder() string {
+	return l.folder
+}
