+++ project/main.go
@@ -1,0 +1,8 @@
+package project
+
+// Project is basically any kind of project (git, local, svn, bzr, nfs...)
+type Project interface {
+	Download() error
+	Update() error
+	Folder() string
+}
